Update object and property names to be more open to future expansion
